refactor: moved VaultPartView to separate file
